pyproject.toml
[project]
name = "labfile-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-google-genai>=3.0.0",
    "python-dotenv>=1.1.1",
]
